review and refactoring

+++ chapter7/projects/13/13.c
@@ -1,7 +1,6 @@
 /*
  * C programming: A Modern Approach. Chapter 07, Project 13
  * Date: February 2023
- * Author: ffstlln
  * Description: Calculates the average word length in a sentence. 
  */
 
@@ -32,6 +31,6 @@ int main(void)
        }
    }
 
-   printf("Average word length: %.1f\n", (double) total_chars / word_count);
+   printf("Average word length: %.1f\n", (float) total_chars / word_count);
    return 0;
 }
